+++ src/pages/components/Newsletter/Newsletter.module.scss
@@ -1,6 +1,7 @@
 @import "../../../styles/mixins/padding.scss";
 @import "../../../styles/mixins/fonts.scss";
 @import "../../../styles/mixins/colors.scss";
+@import "../../../styles/mixins/size.scss";
 
 .main {
       background-color: $darkBlue-color;
@@ -31,13 +32,8 @@
       }
 
       input[type=email] {
-            height: 3rem;
-            border: none;
-            border-radius: 5px;
-            padding: 0 1rem;
+            @include Size;
             margin: 1rem 0;
-            font-size: 1.5rem;
-            font-weight: 400;
             color: black;
             background-color: #dfdfdf;
             transition: all 0.3s ease;
@@ -45,7 +41,7 @@
             &:hover,
             &:focus {
                   transform: scale(1.05);
-                  box-shadow: 0 0 10px #faaa51;
+                  box-shadow: 0 0 10px $gear-color;
                   outline: none;
             }
 
@@ -55,12 +51,7 @@
       }
 
       input[type=submit] {
-            height: 3rem;
-            border: none;
-            border-radius: 5px;
-            padding: 0 1rem;
-            font-size: 1.5rem;
-            font-weight: 400;
+            @include Size;
             background-color: $white-color;
 
             cursor: pointer;
@@ -77,7 +68,7 @@
       @media (max-width: 920px) {
             grid-template-columns: 1fr;
             grid-template-rows: auto auto;
-            
+
             h1 {
                   text-align: center;
                   margin: 1rem 0;
