+++ src/pages/components/button/index.jsx
@@ -1,11 +1,9 @@
 import Styles from "./button.module.scss";
-// provisorio para testar o botao
 
 const Button = ({ title, kind, onClick, type, disabled }) => {
   const generationClassByKind = () => {
     if (kind === "secundary") return Styles.secundary;
     if (kind === "full") return Styles.full;
-
     return Styles.primary;
   };
