+++ src/pages/components/contato/contato.module.scss
@@ -1,6 +1,7 @@
 @import "../../../styles/mixins/fonts.scss";
 @import "../../../styles/mixins/padding.scss";
 @import "../../../styles/mixins/colors.scss";
+@import "../../../styles/mixins/size.scss";
 
 .wrapper {
   display: grid;
@@ -11,7 +12,7 @@
     padding: 2rem;
   }
 
-  @media(max-width: 600px){
+  @media(max-width: 600px) {
     display: block;
   }
 }
@@ -36,7 +37,7 @@
 
   }
 
-  @media(max-width: 600px){
+  @media(max-width: 600px) {
     font-size: small;
   }
 
@@ -72,7 +73,8 @@
     align-items: center;
     text-align: left;
     padding: 0;
-    span{
+
+    span {
       font-size: 1rem;
     }
 
@@ -80,14 +82,12 @@
       font-size: 1.5rem;
     }
 
-    p{
+    p {
       font-size: 1rem;
     }
   }
 
-  @media(max-width: 600px){
-
-  }
+  @media(max-width: 600px) {}
 }
 
 .forms {
@@ -106,10 +106,10 @@
     padding: 0 0 2rem 0;
     margin: 0;
 
-    @media(max-width: 600px){
-    font-size: 1.5rem;
-    padding: 0;
-  }
+    @media(max-width: 600px) {
+      font-size: 1.5rem;
+      padding: 0;
+    }
   }
 
 
@@ -120,30 +120,10 @@
     width: 100%;
     margin: 2rem 0;
 
-
-   /* textarea {
-      width: 100%;
-      height: 10rem;
-      border: 1px solid $accent-color;
-      border-radius: 5px;
-      padding: 1rem;
-      margin: 1rem 0;
-      font-size: 1.5rem;
-      font-weight: 400;
-      color: black;
-    } */
-
-
-
     button {
+      @include Size;
       width: 100%;
-      height: 3rem;
-      border: none;
-      border-radius: 5px;
-      padding: 0 1rem;
       margin: 1rem 0;
-      font-size: 1.5rem;
-      font-weight: 400;
       cursor: pointer;
       transition: all 0.3s ease-in-out;
 
@@ -152,10 +132,10 @@
         color: $white-color;
       }
     }
-    
-    @media(max-width: 600px){
-    margin: 0;
-  }
+
+    @media(max-width: 600px) {
+      margin: 0;
+    }
   }
 
   @media (max-width: 1051px) {
@@ -165,7 +145,7 @@
     width: min-content;
   }
 
-  @media(max-width: 600px){
+  @media(max-width: 600px) {
     width: 15rem;
   }
 
@@ -180,4 +160,3 @@
     display: none;
   }
 }
-
