+++ src/pages/components/features/feature.module.scss
@@ -26,12 +26,14 @@ $font_weight: 300;
             font-weight: $font_weight;
             margin: 0;
       }
+
       @media (max-width: 900px) {
             width: 25rem;
-      } 
+      }
+
       @media (max-width: 600px) {
             width: 18rem;
-      }      
+      }
 }
 
 .cardHolder {
@@ -77,6 +79,6 @@ $font_weight: 300;
 }
 
 .card:hover {
-      box-shadow: 0px 0px 50px 0px #faaa51;
+      box-shadow: 0px 0px 50px 0px $gear-color;
       transform: scale(1.1);
 }
