--- src/pages/components/header/botton/button.module.scss
@@ -1,28 +1,0 @@
-@import "src/styles/mixins/colors.scss";
-@import "src/styles/mixins/fonts.scss";
-
-.button {
-
-    font-weight: bold;
-    border: none;
-    padding: 1.5em 3em;
-    border-radius: 5px;
-    font-family: $font-family-roboto;
-
-    &.primary {
-        background-color: $blue-color;
-        color: $white-color;
-    }
-
-    &.secundary {
-        background: #00171F; // qual a cor?
-        color: $white-color;
-    }
-
-    &:hover {
-        background-color: $accent-color;
-    }
-
-    transition: 0.5s cubic-bezier(0.42, 0, 0.08, 0.96);
-    
-}
