+++ src/pages/components/header/index.jsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Link from "next/link";
 import Logo from "/public/images/logo.svg";
 import Styles from "./header.module.scss";
-import Button from "./botton";
+import Button from "../button";
 import menuIcon from "/public/icons/menu-outline.svg";
 import { useEffect, useState } from "react";
 
@@ -17,7 +17,7 @@ function closeMenu() {
 
 function toggleMenu() {
   var menu = document.getElementById("menu");
-  
+
   if (isOpen) {
     closeMenu();
   } else {
@@ -37,7 +37,7 @@ const Header = () => {
           <Link href="#footer">Home</Link>
           <Link href="https://google.com">O que fazemos</Link>
           <Link href="https://google.com">Cases</Link>
-          <Button title="Fale conosco" onClick={()=>{location.href='#form'}}/>
+          <Button title="Fale conosco" onClick={() => { location.href = '#form' }} />
         </div>
         <Image className={Styles.icon} src={menuIcon} onClick={toggleMenu} />
       </div>
