+++ src/pages/components/input/input.module.scss
@@ -1,26 +1,22 @@
 @import "src/styles/mixins/colors.scss";
+@import "src/styles/mixins/size.scss";
 
 .input {
+  @include Size;
   width: 25rem;
-  height: 3rem;
-  border: none;
-  border-radius: 5px;
-  padding: 0 1rem;
   margin: 1rem 0;
-  font-size: 1.5rem;
-  font-weight: 400;
   color: black;
   background-color: #dfdfdf;
   transition: all 0.3s ease;
 
   &:hover,
-  &:focus { 
-    transform: scale(1.05); 
-    box-shadow: 0 0 10px #faaa51;
+  &:focus {
+    transform: scale(1.05);
+    box-shadow: 0 0 10px $gear-color;
     outline: none;
   }
 
-  @media(max-width: 600px){
+  @media(max-width: 600px) {
     font-size: 1rem;
     width: 13rem;
   }
