+++ src/pages/components/select/select.module.scss
@@ -1,14 +1,10 @@
 @import "src/styles/mixins/colors.scss";
+@import "src/styles/mixins/size.scss";
 
 .select {
+  @include Size;
   width: 27rem;
-  height: 3rem;
-  border: none;
-  border-radius: 5px;
-  padding: 0 1rem;
   margin: 1rem 0;
-  font-size: 1.5rem;
-  font-weight: 400;
   color: black;
   background-color: #dfdfdf;
