+++ src/pages/index.jsx
@@ -8,8 +8,6 @@ import Newsletter from './components/Newsletter'
 
 export default function Home() {
 
-  
-
   return (
     <div className={styles.description}>
       <Header />
