+++ src/styles/mixins/colors.scss
@@ -5,3 +5,4 @@ $accent-color: #00A8E8;
 $whitesmoke-color: #F5F5F5;
 $darkBlue-color: #137aaa48;
 $gradientBlue-color: linear-gradient(#407193, $blue-color);
+$gear-color: #faaa51;
