+++ src/styles/mixins/fonts.scss
@@ -1,4 +1,3 @@
-
 $font-family: Arial, Helvetica, sans-serif;
 
 $font-family-roboto: 'Roboto', monospace;
