+++ src/styles/mixins/size.scss
@@ -1,0 +1,8 @@
+@mixin Size {
+    height: 3rem;
+    border: none;
+    border-radius: 5px;
+    padding: 0 1rem;
+    font-size: 1.5rem;
+    font-weight: 400;
+}
